requirements.txt
streamlit
pandas
plotly.express
plotly.graph_objects
kiteconnect